requirements.txt
asgiref==3.5.2
Django==4.1.3
django-admin-interface==0.23.0
django-colorfield==0.7.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-htmx==1.13.0
djangorestframework==3.14.0
Pillow==9.3.0
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
